FeatureAnalysis

a function can generate mz vs rt plot and QC distribution plot, also can filter isotope,rsd and zero value.

FeatureAnalysis(
  zero.filter = FALSE,
  RSD.filter = FALSE,
  zero.check = TRUE,
  mzrt = FALSE,
  FilterIsotope = TRUE
)

Arguments

zero.filter
default is FALSE,if the zero value exsit in your data, make FALSE to TRUE.

RSD.filter
default is FALSE,if the percentage of qc rsd larger than 0.3 more than 0.7,almost after normalization,make FALSE to TRUE.

zero.check
default is TRUE.

mzrt
default is FALSE.

FilterIsotope
default is TRUE.

Value

All the results can be got form other functions and instruction.

Examples

# \donttest{
##---- Be sure the format of data and sample.info is correct!! ----
##create a folder for Shine demo
dir.create("Demo for Shine")

#> Warning: 'Demo for Shine'已存在

setwd("Demo for Shine")
#load the demo data
data(data, package = "Shine")

#> Warning: data set ‘data’ not found

data(sample.info, package = "Shine")

#> Warning: data set ‘sample.info’ not found

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)

#> Error in as.data.frame.default(x[[i]], optional = TRUE): cannot coerce class ‘"function"’ to a data.frame

write.csv(sample.info, "sample.info.csv", row.names = FALSE)

#> Error in is.data.frame(x): 找不到对象'sample.info'

#run FeatureAnalysis
FeatureAnalysis(data = data,sample.info = sample.info,RSD.filter = T,
zero.check = F)

#> Error in FeatureAnalysis(data = data, sample.info = sample.info, RSD.filter = T,     zero.check = F): 参数没有用(data = data, sample.info = sample.info)

# }